+++ laktory/resources/quickstart-stacks/workflows/notebooks/jobs/job_laktory_pl.py
@@ -1,5 +1,4 @@
 # COMMAND ----------
-dbutils.widgets.text("workspace_laktory_root", "/.laktory/")
 dbutils.widgets.text("pipeline_name", "pl-stocks-job")
 dbutils.widgets.text("node_name", "")
 dbutils.widgets.text("full_refresh", "False")